markdown_helpdesk_backend/rag_service/requirements.txt
fastapi
uvicorn
pyjwt
langchain
faiss-cpu
chromadb